+++ src/Beef.Core/NotFoundException.cs
@@ -11,6 +11,8 @@ namespace Beef
     /// <remarks>The <see cref="Exception.Message"/> defaults to: <i>Requested data was not found.</i></remarks>
     public class NotFoundException : Exception, IBusinessException
     {
+        private const string _message = "Requested data was not found.";
+
         /// <summary>
         /// Get or sets the <see cref="ShouldBeLogged"/> value.
         /// </summary>
@@ -19,52 +21,34 @@ namespace Beef
         /// <summary>
         /// Initializes a new instance of the <see cref="NotFoundException"/> class.
         /// </summary>
-        public NotFoundException()
-            : base(new LText("Beef.NotFoundException"))
-        {
-        }
+        public NotFoundException() : base(new LText("Beef.NotFoundException")) { }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="NotFoundException"/> class with a specified messsage.
         /// </summary>
         /// <param name="message">The message text.</param>
-        public NotFoundException(string? message)
-            : base(message ?? new LText("Beef.NotFoundException"))
-        {
-        }
+        public NotFoundException(string? message) : base(message ?? new LText("Beef.NotFoundException", _message)) { }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="NotFoundException"/> class with a specified messsage and inner exception.
         /// </summary>
         /// <param name="message">The message text.</param>
         /// <param name="innerException">The inner <see cref="Exception"/>.</param>
-        public NotFoundException(string? message, Exception innerException)
-            : base(message ?? new LText("Beef.NotFoundException"), innerException)
-        {
-        }
+        public NotFoundException(string? message, Exception innerException) : base(message ?? new LText("Beef.NotFoundException", _message), innerException) { }
 
         /// <summary>
         /// Gets the <see cref="ErrorType"/> (see <see cref="ErrorType.NotFoundError"/>).
         /// </summary>
-        public ErrorType ErrorType
-        {
-            get { return ErrorType.NotFoundError; }
-        }
+        public ErrorType ErrorType => ErrorType.NotFoundError; 
 
         /// <summary>
         /// Gets the corresponding <see cref="HttpStatusCode"/>.
         /// </summary>
-        public HttpStatusCode StatusCode
-        {
-            get { return HttpStatusCode.NotFound; }
-        }
+        public HttpStatusCode StatusCode => HttpStatusCode.NotFound; 
 
         /// <summary>
         /// Indicates whether the <see cref="Exception"/> should be logged (returns the <see cref="ShouldExceptionBeLogged"/> value).
         /// </summary>
-        public bool ShouldBeLogged
-        {
-            get { return ShouldExceptionBeLogged; }
-        }
+        public bool ShouldBeLogged => ShouldExceptionBeLogged; 
     }
 }